SYWEB-116: Implement historical display name support.
This works for both live and paginated events. Each 'message' event has
an associated '__room_member' key which points to the state of the sender
at that point in time. Invites have an additional key '__target_room_member'
which point to the state of the invitee at that point in time. This obviates
the need for mapping user_ids to *current* displaynames in the message list,
though this is still required for the user/presence list.