Alter phrasing to clarify where info is stored.
A user on #matrix:matrix.org was confused by the phrasing of the first
sentence in the paragraph and couldn't tell whether it was saying that
the homeserver stored the data or the clients did. This change splits it
into two sentences to make the subject of each sentence clear.