Use a special matrix.to URL for room notifs
So you'll have to explicitly accept the pill completion to actually
do a room notification.

Note that this matches on formatted_body as the URL will only be
in the formatted_body. It could also match for only HTML bodies,
but the URL ought to be specific enough that it's sensible to match
whatever the format is?