docker: Allow customization of TLS cert/key/DH and signing key paths
I'm starting to move an existing Synapse installation into Kubernetes,
and it would be easier to manage if I could customize the paths for
Synapse's TLS cert/key/etc. Kubernetes allows you to mount a
`Secret` (possibly containing multiple files) as a volume in your
containers, but you can't mount one of these volumes on top of another
persistent volume.

Signed-off-by: Jonathan Frederickson <jonathan@terracrypt.net>