Add concept of StatelessContext
This allows us to lazily load state on the master process as required
(when using workers). The advantage of this is that in the common case
the master doesn't need the full state of the events to persist them.

We keep the concept of an EventContext that has the current_state_ids
and prev_state_ids attributes, to avoid having to rewrite a lot of code.
Instead, we shift the necessary functions to accept StatelessContext.